langchain_projects/requirements.txt
langchain_openai 
langchain_core
python-dotenv
streamlit
langchain_community
langserve
fastapi
uvicorn
sse_starlette
ipykernel
bs4
pypdf
langchain
chromadb
faiss-cpu
arxiv
wikipedia
langchainhub
groq
langchain_groq